requirements.txt
streamlit
streamlit-chat
streamlit-webrtc
cohere
python-dotenv
PyPDF2
pydub
SpeechRecognition
